[FIX] mail: cannot route catchall bounced email
Configure a catchall address in Odoo, send an email to that very
address, it should bounce back to the user asking him not to send any
email to the catchall but the email is never sent.

Internally, Odoo reply to the user using the `To` header of the original
email as `From` of the bounced email. It should never use the catchall
address as it is not a valid email to send email from.

closes odoo/odoo#53036

Task: 2277661
X-original-commit: d870f67cb7be4bdcb135bdad3a1fa489777a3f31
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>