[FIX] mrp: forgotten create decorator
Followup of
https://github.com/odoo/odoo/commit/bd72217a7c6b3856fd793f7e3054f7c26d2f48ce.

By default, the create, when not specified as `api.model`, supports
creation in batch (thus the batching opportunity isn't lost here).

BUT, if a create override isn't specified as `api.model` or
`api.model_create_multi`, it won't be accessible in rpc calls (because
the ORM thinks a record in self is needed).

It doesn't look like somebody can create stock move lines from the web
client itself, but it doesn't cost anything to follow the global api
(and it's clearer to clearly specify the create as create_multi).

closes odoo/odoo#48964

X-original-commit: 5f62483e0871ea9bd3555ed745b685f045e61d81
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>